refactor: API 및 DTO 클래스 패키지 분리 (#133)
