--- src/main/java/site/gachontable/gachontablebe/domain/admin/presentation/dto/request/AdminLoginRequest.java
@@ -1,5 +1,0 @@
-package site.gachontable.gachontablebe.domain.admin.presentation.dto.request;
-
-public record AdminLoginRequest(String id,
-                                String password) {
-}
