--- src/main/java/site/gachontable/gachontablebe/domain/admin/presentation/dto/request/AdminRegisterRequest.java
@@ -1,7 +1,0 @@
-package site.gachontable.gachontablebe.domain.admin.presentation.dto.request;
-
-public record AdminRegisterRequest(String username,
-                                   String password,
-                                   String tel,
-                                   Integer pubId) {
-}
