--- src/main/java/site/gachontable/gachontablebe/domain/admin/presentation/dto/request/EnterUserRequest.java
@@ -1,7 +1,0 @@
-package site.gachontable.gachontablebe.domain.admin.presentation.dto.request;
-
-import java.util.UUID;
-
-public record EnterUserRequest(UUID waitingId,
-                               Integer seatingNum) {
-}
