--- src/main/java/site/gachontable/gachontablebe/domain/admin/presentation/dto/request/PubManageRequest.java
@@ -1,13 +1,0 @@
-package site.gachontable.gachontablebe.domain.admin.presentation.dto.request;
-
-import java.util.List;
-
-public record PubManageRequest(List<String> thumbnails,
-                               List<MenuRequest> menuRequests) {
-    public record MenuRequest(Integer menuId,
-                              String thumbnail,
-                              String menuName,
-                              String price,
-                              String oneLiner) {
-    }
-}
