--- src/main/java/site/gachontable/gachontablebe/domain/admin/presentation/dto/response/AdminLoginResponse.java
@@ -1,6 +1,0 @@
-package site.gachontable.gachontablebe.domain.admin.presentation.dto.response;
-
-public record AdminLoginResponse(String accessToken,
-                                 String refreshToken,
-                                 Integer pubId) {
-}
