--- src/main/java/site/gachontable/gachontablebe/domain/auth/presentation/dto/response/AuthResponse.java
@@ -1,6 +1,0 @@
-package site.gachontable.gachontablebe.domain.auth.presentation.dto.response;
-
-public record AuthResponse(String accessToken,
-                           String refreshToken,
-                           String username) {
-}
