--- src/main/java/site/gachontable/gachontablebe/domain/user/presentation/dto/request/UserLoginRequest.java
@@ -1,5 +1,0 @@
-package site.gachontable.gachontablebe.domain.user.presentation.dto.request;
-
-public record UserLoginRequest(String id,
-                               String password) {
-}
