--- src/main/java/site/gachontable/gachontablebe/domain/waiting/presentation/dto/response/WaitingResponse.java
@@ -1,5 +1,0 @@
-package site.gachontable.gachontablebe.domain.waiting.presentation.dto.response;
-
-public record WaitingResponse(Boolean status,
-                              String message) {
-}
