+++ src/main/java/site/gachontable/presentation/admin/dto/request/AdminRegisterRequest.java
@@ -1,0 +1,7 @@
+package site.gachontable.presentation.admin.dto.request;
+
+public record AdminRegisterRequest(String username,
+                                   String password,
+                                   String tel,
+                                   Integer pubId) {
+}
