+++ src/main/java/site/gachontable/presentation/admin/dto/request/CallUserRequest.java
@@ -1,0 +1,6 @@
+package site.gachontable.presentation.admin.dto.request;
+
+import java.util.UUID;
+
+public record CallUserRequest(UUID waitingId) {
+}
