+++ src/main/java/site/gachontable/presentation/admin/dto/request/EnterUserRequest.java
@@ -1,0 +1,7 @@
+package site.gachontable.presentation.admin.dto.request;
+
+import java.util.UUID;
+
+public record EnterUserRequest(UUID waitingId,
+                               Integer seatingNum) {
+}
