+++ src/main/java/site/gachontable/presentation/admin/dto/request/UpdateStatusRequest.java
@@ -1,0 +1,4 @@
+package site.gachontable.presentation.admin.dto.request;
+
+public record UpdateStatusRequest(Boolean status) {
+}
