+++ src/main/java/site/gachontable/presentation/admin/dto/response/AdminLoginResponse.java
@@ -1,0 +1,6 @@
+package site.gachontable.presentation.admin.dto.response;
+
+public record AdminLoginResponse(String accessToken,
+                                 String refreshToken,
+                                 Integer pubId) {
+}
