+++ src/main/java/site/gachontable/presentation/auth/dto/response/AuthResponse.java
@@ -1,0 +1,6 @@
+package site.gachontable.presentation.auth.dto.response;
+
+public record AuthResponse(String accessToken,
+                           String refreshToken,
+                           String username) {
+}
