+++ src/main/java/site/gachontable/presentation/member/dto/request/UserLoginRequest.java
@@ -1,0 +1,5 @@
+package site.gachontable.presentation.member.dto.request;
+
+public record UserLoginRequest(String id,
+                               String password) {
+}
