+++ src/main/java/site/gachontable/presentation/pub/dto/request/PubRegisterRequest.java
@@ -1,0 +1,10 @@
+package site.gachontable.presentation.pub.dto.request;
+
+public record PubRegisterRequest(String pubName,
+                                 String oneLiner,
+                                 String instagramUrl,
+                                 Integer minutes,
+                                 String menuUrl,
+                                 Boolean openStatus,
+                                 Boolean waitingStatus) {
+}
