+++ src/main/java/site/gachontable/presentation/waiting/dto/request/CancelRequest.java
@@ -1,0 +1,6 @@
+package site.gachontable.presentation.waiting.dto.request;
+
+import java.util.UUID;
+
+public record CancelRequest(UUID waitingId) {
+}
