+++ src/main/java/site/gachontable/presentation/waiting/dto/request/RemoteWaitingRequest.java
@@ -1,0 +1,7 @@
+package site.gachontable.presentation.waiting.dto.request;
+
+import site.gachontable.domain.shared.Table;
+
+public record RemoteWaitingRequest(Integer pubId,
+                                   Table tableType) {
+}
