+++ src/main/java/site/gachontable/presentation/waiting/dto/response/WaitingResponse.java
@@ -1,0 +1,5 @@
+package site.gachontable.presentation.waiting.dto.response;
+
+public record WaitingResponse(Boolean status,
+                              String message) {
+}
